JC constants-related changes
+ add test for checking conformance of custom constants to JC
standard-defined constants (ConstsNGTest::verifyConstantsValidity). It
is verified to be valid now and commented out at the moment (for proper
check, latest version of JC library must be included in project. But
this will colide with jcardsim simulator that supports only 2.2.2 and
cause exception when simulator is run as implementaton is taken from JC
library, not jcardsim)
- removed RSA 3072 (not defined in standard)
- fixed improper name LENGTH_HMAC_SHA_384_BLOCK_128